+++ index.js
@@ -1,4 +1,5 @@
 const EventEmitter = require('events');
+const merkle = require('merkle-tools')
 
 class APP extends EventEmitter {
 
@@ -10,6 +11,7 @@ class APP extends EventEmitter {
 
         if (!this.PEER)
             this.PEER = require("./lib/entries/peer")(this);
+
         if (!this.DATA)
             this.DATA = require("./lib/entries/data")(this);
 
@@ -22,12 +24,18 @@ class APP extends EventEmitter {
         if (!this.CONSENSUS)
             this.CONSENSUS = require("./lib/entries/consensus")(this);
 
+        if (!this.VALIDATOR)
+            this.VALIDATOR = require("./lib/entries/validator")(this);
+
+        if (!this.ROUNDMANAGER)
+            this.ROUNDMANAGER = require("./lib/mappers/round")(this);
+
         this.CONSENSUS.Centralized = require('./algo/centralized')(this);
         this.CONSENSUS.ProofOfWorkConsensus = require('./algo/pow')(this);
         this.CONSENSUS.ProofOfStakeConsensus = require('./algo/pos')(this);
         this.CONSENSUS.DelegatedProofOfWorkConsensus = require('./algo/static-dpow')(this);
         this.CONSENSUS.DelegatedProofOfStakeConsensus = require('./algo/static-dpos')(this);
-
+        this.CONSENSUS.DynamicDelegateProofOfStakeConsensus = require('./algo/dynamic-dpos')(this);
     }
     definePeerClass(man) {
         this.PEER = man;
@@ -50,6 +58,18 @@ class APP extends EventEmitter {
     defineConsensusClass(man) {
         this.CONSENSUS = man;
     }
+    defineValidatorClass(man) {
+        this.VALIDATOR = man;
+    }
+    defineRoundManagerClass(man) {
+        this.ROUNDMANAGER = man;
+    }
+    setPeerManager(man) {
+        this.peerManager = man;
+    }
+    setRoundManager(man) {
+        this.roundManager = man;
+    }
     start(consensus) {
 
         this.emit("app.config", this.config);
@@ -60,6 +80,9 @@ class APP extends EventEmitter {
         if (!this.dataManager)
             this.dataManager = new this.DATAMANAGER();
         this.emit("app.datamanager");
+        if (!this.roundManager)
+            this.roundManager = new this.ROUNDMANAGER();
+        this.emit("app.roundmanager");
 
         //if (!this.config.algorithm)
         //    throw new Error('Algorithm of consensus must be defined');
@@ -76,14 +99,16 @@ class APP extends EventEmitter {
             this.consensus = new (algoritm_fnc(this));
         } else {
             let cls = null;
+            if (algorithm_name == 'ddpos')
+                cls = this.CONSENSUS.DynamicDelegateProofOfStakeConsensus;
             if (algorithm_name == 'dpow')
                 cls = this.CONSENSUS.DelegatedProofOfWorkConsensus;
             if (algorithm_name == 'dpos')
-                cls = this.CONSENSUS.DelegatedProofOfStakeConsensus;
+                cls = this.CONSENSUS.DelegatedProofOfStakeConsensus;//dpow+pos
             if (algorithm_name == 'pow')
                 cls = this.CONSENSUS.ProofOfWorkConsensus;
             if (algorithm_name == 'pos')
-                cls = this.CONSENSUS.ProofOfStakeConsensus;
+                cls = this.CONSENSUS.ProofOfStakeConsensus;//pow+pos
             if (algorithm_name == 'centralized')
                 cls = this.CONSENSUS.Centralized;
 
@@ -97,7 +122,7 @@ class APP extends EventEmitter {
         this.emit("app.selected_consensus", this.consensus_name);
         this.consensus.init();
     }
-    getConsensus(){
+    getConsensus() {
         return this.consensus;
     }
     getDefaultConfig() {
@@ -132,6 +157,15 @@ class APP extends EventEmitter {
                 'delegateMode': true,
                 'delegates': []//if this param is empty - we can make dynamic delegates
             },
+            'ddpos': {
+                'extends': 'dpos',
+                'validatorCount': 60,
+                'timeout': 60,//timeout in seconds for sending block for validator. If timedout - decrement priority and set cursor to next
+                'staticDelegatesLimit': 5,//enable static delegates from config if connected validator count less then this parameter
+                'delegates': [],//if this param is empty - we can make dynamic delegates
+                "timeout": 60, //max block time after prev block
+                "pause": 20,//min block time after prev block
+            },
             "genesis": { //need to be rediclared on yours config
                 "id": 'genesis',
                 "prev": -1,
@@ -140,6 +174,29 @@ class APP extends EventEmitter {
                 "nonce": 0,
             }
         };
+    }
+    merkle(list) {
+
+        function reverseBuffer(buff) {
+            var out_rev = Buffer.alloc(buff.length), i = 0
+            for (; i < buff.length; i++) {
+                out_rev[buff.length - 1 - i] = buff[i];
+            }
+
+            return out_rev;
+        }
+
+        function makeMerkle(arr) {
+            let m = new merkle()
+            for (let i in arr) {
+                m.addLeaf(reverseBuffer(new Buffer(arr[i].replace('0x', ''), 'hex')).toString('hex'))
+            }
+
+            m.makeBTCTree(true);
+            return reverseBuffer(new Buffer(m.getMerkleRoot(), 'hex')).toString('hex')
+        }
+
+        return makeMerkle(list);
     }
 
 }